# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для биллинга",
        "description": "Этот API предоставляет систему биллинга, в которой пользователи могут зарегистрироваться, проверять свой баланс, выполнять транзакции (пополнение/снятие) и просматривать историю транзакций. Все операции проходят проверку, исключая возможность отрицательного баланса.",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Локальный сервер для разработки API биллинга."
        }
    ],
    "paths": {
...
        "/api/user/transaction/deposit": {
            "post": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Пополнение баланса",
                "description": "Этот метод позволяет пользователю отправить запрос на пополнение баланса. Транзакция обрабатывается в фоновом режиме.",
                "operationId": "64238b3c8935b38e8d11a5e604e14e7d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Сумма пополнения (минимум 1)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос на снятие успешно отправлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Your diposit request is being processed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string",
                                            "example": "Поле amount является обязательным."
                                        },
                                        "errors": {
                                            "description": "Детализированные ошибки валидации",
                                            "type": "object",
                                            "example": {
                                                "amount": [
                                                    "Поле amount является обязательным."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке аутентификации",
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/transaction/withdraw": {
            "post": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "Снятие средств",
                "description": "Этот метод позволяет пользователю отправить запрос на снятие средств. Транзакция обрабатывается в фоновом режиме.",
                "operationId": "d1884a7aacc81dd693cc624e72e75a63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "amount"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Сумма пополнения (минимум 1)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Запрос на снятие успешно отправлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Your withdraw request is being processed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string",
                                            "example": "Поле amount является обязательным."
                                        },
                                        "errors": {
                                            "description": "Детализированные ошибки валидации",
                                            "type": "object",
                                            "example": {
                                                "amount": [
                                                    "Поле amount является обязательным."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке аутентификации",
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user/transaction/history": {
            "get": {
                "tags": [
                    "Транзакции"
                ],
                "summary": "История транзакций",
                "description": "Этот метод позволяет пользователю получить историю транзакций, с возможностью указания ограничения на количество записей.",
                "operationId": "59235997e6b2a8aa4b0aa8942ec37fc9",
                "parameters": [
                    {
                        "name": "page_limit",
                        "in": "query",
                        "description": "Ограничение на количество транзакций на странице (максимум 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "История транзакций успешно получена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Флаг успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "content": {
                                            "description": "Список транзакций",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "type": {
                                                        "description": "Тип транзакции (например, депозит)",
                                                        "type": "string",
                                                        "example": "deposit"
                                                    },
                                                    "amount": {
                                                        "description": "Сумма транзакции",
                                                        "type": "string",
                                                        "example": "100.00"
                                                    },
                                                    "description": {
                                                        "description": "Описание транзакции",
                                                        "type": "string",
                                                        "example": "Deposit successful"
                                                    },
                                                    "created_at": {
                                                        "description": "Дата и время создания",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-31T18:32:08.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "description": "Дата и время последнего обновления",
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2025-01-31T18:32:08.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Описание ошибки",
                                            "type": "string",
                                            "example": "Поле page_limit должно быть числом."
                                        },
                                        "errors": {
                                            "description": "Детализированные ошибки валидации",
                                            "type": "object",
                                            "example": {
                                                "page_limit": [
                                                    "Поле page_limit должно быть числом."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке аутентификации",
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Используйте Bearer Token для аутентификации. Все защищенные маршруты требуют действительного токена после входа пользователя в систему.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}